Extract AutoScrollHandler
